move enum into data folder
